requirements.txt
streamlit
python-dotenv
sentence-transformers
faiss-cpu
PyPDF2
requests
